Create a `struct` called `Tags` to hold the tags
Should make it a bit easier to reason about what the values are in
`Tags` and how they should be handled. Fields are ordered to match the
index order in `uint64_t` arrays that hold these tag values. This is
done primarily as a convenience to the developer and is not necessarily
guaranteed to be kept by the compiler.