Mark C level functions as `nogil`
These functions should be safe to use without the Python GIL as they are
not creating or destroying Python objects. So mark them as such.

Note: This does **not** actually release the GIL. We still have to do
that ourselves. However this allows us to use these in code sections
where we have already released the GIL. We can also continue to use them
when we have the GIL as well.