Make callbacks `nogil`, but acquire GIL internally
As these functions are effectively called from `nogil` contexts now
given the C functions calling them are `nogil`, make them `nogil` as
well, but acquire the GIL internally for the parts of the function that
require it. Namely when the Python callback is run and its results are
added to the asyncio Future.