Just use `memoryview` for NumPy as well
NumPy is the main user of `__array_interface__`. However NumPy also
supports the Python Buffer Protocol, which is what is typically used for
this purpose. This can easily be used with `memoryview`s as we have done
here. So simply drop the `__array_interface__` check and use
`memoryview` for NumPy as well.