Cast `iface["data"]` as `tuple`
By default Cython generates a bunch of code to handle `tuple`s, `list`s,
and generally iterable objects in Python for doing this unpacking. As we
know we have a `tuple` and can thus simplify the code here a bit, go
ahead and cast `iface["data"]` as a `tuple`. This way it can focus on
the code needed to unpack a `tuple` and drop everything else extra.