Use `struct` with data (and it's constituents)
Since the `__cuda_array_interface__` (and for that matter the
`__array_interface__`) treat the `data` value as a `tuple` containing
both the pointer and whether the data is readonly or not, update our
code to treat it similarly at the Cython level by using a `struct` for
it. This also results in a bit more efficient code.