Make sure type `str`'s in `dict` are `intern`ed
Lookups in Python `dict`s with `str` keys can be sped up if the keys are
`intern`ed and the key accessed is `intern`ed. We use this to `intern`
all of the keys of `itemsize_mapping` to speedup lookup times with this
`dict`. This way if we encounter a type `str`, which is `intern`ed
Python can speedup lookup time a bit.