Type `send`, `recv` arguments and `port` (#612)
* Type `port` as `uint16_t`

As C libraries we interface with use `uint16_t` to type port variables,
make sure we do the same.

* Type `*args` and `**kwargs` as `tuple` and `dict`

* Type `name` as `str`

* Use `None` for default arguments

As some objects like `dict()` are mutable, these are generally not used
as default arguments and are set within functions. To fix this, we set
these to `None` instead and then replace those values internally if they
are `None`.

* Use `()` and `{}` for empty `tuple` and `dict`

Cython is able to pick up on this particular syntax a bit better and as
result optimize it a bit better. Should avoid some overhead compared to
`tuple()` and `dict()`, which may turn into Python function calls
instead.

* Type `cb_args` & `cb_kwargs` as `tuple` & `dict`