Use C attributes, methods, & minor error message cleanup (#629)
* Use `UCXRequest`'s `_handle` directly

* Grab `UCXRequest`'s `info` via `_handle`

This is able to proceed efficiently directly through C without needing
to go through Python. Also behaves equivalent to how the `info` property
would behave. Include the `.closed()` check that we would otherwise use
in the property for good measure.

* Combine definition and assignment on one line

* Provide efficient C-level access to `closed()`

In addition to exposing `UCXRequest.closed()` to Python, expose an
efficient version to C as well. Also type its `return` value as `bint`
to provide an efficient C type we can use. As no exceptions can be
raised from this function skip marking it as `except *`.

* Provide C implementation of `.close()` too

Adds a C-level implementation of `.close()` for `UCXRequest` as well.

* Fix tense

* Construct error message only when needed

* Add more methods to C API

As we are calling these methods from Cython, make sure we expose a C API
for faster calling. This done in addition to the Python API that was
already available.