Fix typing of `ucp_listener_h`
As `ucp_listener_h` is actually a pointer and not a `struct`, correct
the typing of `ucp_listener_h` to reflect this. Also as the underlying
type is `ucp_listener`, make sure this is also added to our definitions.
This ensures that this part of the casting in
`_ucx_listener_handle_finalizer` is not dropped by Cython.

https://github.com/openucx/ucx/blob/8b6641b02ef5d217323f719d54dfb2e9142dad59/src/ucp/api/ucp_def.h#L168-L175