Fix compile warnings (#641)
* Cast `handle` to `void*` first

This should make sure Cython recognizes this is a pointer and thus keep
the casting as specified when generating C code. As a result the C code
should still have the cast and thus avoid the warning regarding pass an
integer as a pointer.

* Wrap `_get_error_callback` arguments on next line

* Note `_get_error_callback` can raise exceptions

* Fix typing of `ucp_err_handler_cb_t`

As this is a function pointer and not a `struct`, correct its typing
based on its upstream definition.

https://github.com/openucx/ucx/blob/8b6641b02ef5d217323f719d54dfb2e9142dad59/src/ucp/api/ucp_def.h#L335-L348

* Fix typing of `ucp_listener_h`

As `ucp_listener_h` is actually a pointer and not a `struct`, correct
the typing of `ucp_listener_h` to reflect this. Also as the underlying
type is `ucp_listener`, make sure this is also added to our definitions.
This ensures that this part of the casting in
`_ucx_listener_handle_finalizer` is not dropped by Cython.

https://github.com/openucx/ucx/blob/8b6641b02ef5d217323f719d54dfb2e9142dad59/src/ucp/api/ucp_def.h#L168-L175

* Space out `enum` declaration

* Drop unneeded `void*` cast

Now that `ucp_listener_h` is typed correctly, we no longer need to cast
through a `void*` first. So go ahead and drop that.