Avoid explicitly closing loops in benchmarks/tests
This is an attempt to fix errors such as:

```
sys:1: RuntimeWarning: coroutine 'BlockingMode._arm_worker' was never awaited
RuntimeWarning: Enable tracemalloc to get the object allocation traceback
Task was destroyed but it is pending!
```

This happens because the event loop is closed before the rest of
UCX-Py's finalizers/destructors are executed, which prevents the
`_arm_worker` task from being cancelled in `ProgressTask`'s destructor.