Fix test warnings (#800)
* Remove UCX_PATH from CI build script

UCX_PATH isn't used anymore and because it's not a UCX variable in
itself, it raises warnings:

```
parser.c:1888 UCX  WARN  unused env variable: UCX_PATH (set UCX_WARN_UNUSED_ENV_VARS=n to suppress this warning)
```

* Add bytearray assert function

Using the NumPy is unsafe with empty bytearrays and throws warnings that
will eventually be errors:

```
DeprecationWarning: The truth value of an empty array is ambiguous.  Returning False, but in future this will result in an error. Use `array.size > 0` to check that an array is not empty.
```