Fix core test_close_callback
Since ep.close() calls worker.progress() it can't be called within the
listener callback, therefore we have to progress outside until the
callback is called. We must check the processes' exitcodes to be `0`, if
they timeout the exitcode is `None`, which can't be checked correctly
just with `not process.exitcode`.