Add dockerfile using librdmacm to provide infiniband interfaces
Rather than relying on exact matching MOFED versions, use rdma-core
which is forward-compatible with MOFED version >= 5.x.

To facilitate downstream users adopting the build setup, split the
building and installing of UCX and UCX-Py into scripted steps for which
the conda environment name and UCX version to use can be configured.
Adapt the existing MOFED-based docker file to use this setup too.

Closes #846.