Fix `RAPIDS_VERSION` environment variable
This PR fixes the `RAPIDS_VERSION` environment variable in `ci/gpu/build.sh` that seems to be breaking CI.

Ordinarily, I'd also update the `update-version.sh` script, but this problematic script will be going away when we migrate `ucx-py` to GitHub Actions, which will be happening before the next release.

Therefore it's fine to just manually edit this string for now.