add compilers to `ucx-py` recipe
throughout the process of upgrading the RAPIDS libraries to `11.8`, I've noticed that the `11.8` images are missing a lot of packages that were previously available in the `11.5` images.

these changes are from the upstream images, not RAPIDS' images.

it seems that `gcc` is missing, which is causing build issues for `ucx-py`.

this commit adds conda's `c` and `cxx` compilers to account for these new missing dependencies.