Pin cupy and unset version (#938)
cupy just released version 12, which we do not yet support, so wheels CI must be pinned to require a lower version. We also decided not to overwrite the `__version__` in wheels on other RAPIDS repos, so this PR makes changes to match that.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Sevag H (https://github.com/sevagh)

URL: https://github.com/rapidsai/ucx-py/pull/938