Make final discontinuation changes (#1169)
This is expected to be the last commit to UCX-Py. It updates the README and docs warnings to inform users that UCX-Py has been permanently discontinued. Along with it, this change removes CUDA 13 mentions from docs, since there will be no CUDA 13 packages released for UCX-Py, however, the functional changes to allow building with CUDA 13 are kept should those be needed for any reason in the near-term.

Authors:
  - Peter Andreas Entschev (https://github.com/pentschev)

Approvers:
  - James Lamb (https://github.com/jameslamb)
  - Mads R. B. Kristensen (https://github.com/madsbk)

URL: https://github.com/rapidsai/ucx-py/pull/1169